[System page]Step 1: Added few of the system page functions in a separate file and imported them. Status: Everything works!
